third commit: calc2 can now do division with decimals

+++ Calc2/app/src/main/java/com/example/calc2/MainActivity.java
@@ -24,9 +24,9 @@ public class MainActivity extends AppCompatActivity {
     Button btnMinus;
     Button btnMult;
     Button btnDiv;
-    int result;
-    int firstNum;
-    int secondNum;
+    float result;
+    float firstNum;
+    float secondNum;
     String operator;
     String temp;
